+++ CHANGELOG.md
@@ -5,6 +5,12 @@
   * IntuneDeviceManagementCompliancePartner
     Initial release
 
+* MISC
+  * Fixed the Fabric web request to use basic parsing.
+* M365DSCUtil
+  * Update `Get-M365DSCWorkloadsListFromResourceNames` function for more input types.
+    FIXES [#5525](https://github.com/microsoft/Microsoft365DSC/issues/5525)
+
 # 1.24.1211.1
 
 * AADApplication